Add files via upload
Implemented the LSTM model for product sales forecasting for 3 months in the future. A part of the dataset used can be found here- https://www.kaggle.com/c/demand-forecasting-kernels-only/data.
The dataset contains sales data for 5 years(2013-2017) for 10 stores and 50 items. Due to lack of my computing power I just used the LSTM model for a fraction of the dataset which includes 2 years (2016 & 2017) sales data for 2 stores and 5 items and predicted the sales for the first 3 months of 2018.